+++ Silownia/Silownia/Controllers/TreningPersonalnyController.cs
@@ -22,7 +22,7 @@ namespace Silownia.Controllers
         // GET: TreningPersonalny
         public ActionResult Index(string imieNazwisko, string SilowniaID, string TrenerID, string SpecjalizacjaID, int page = 1, int pageSize = 10, AkcjaEnumTrening akcja = AkcjaEnumTrening.Brak, String info = null)
         {
-          //  if (Session["User"] != null)
+            if (Session["User"] != null)
             {
                 ViewBag.SilowniaID = new SelectList(db.Silownie.DistinctBy(a => new { a.Nazwa }), "Nazwa", "Nazwa");
                 ViewBag.TrenerID = new SelectList(db.Trenerzy.DistinctBy(a => new { a.Pesel }), "imieNazwisko", "imieNazwisko");
@@ -60,13 +60,13 @@ namespace Silownia.Controllers
 
                 return View(model);
             }
-         //   return HttpNotFound();
+            return HttpNotFound();
         }
 
         // GET: TreningPersonalny/Details/5
         public ActionResult Details(int? id)
         {
-           // if (Session["User"] != null)
+            if (Session["User"] != null)
             {
                 if (id == null)
                 {
@@ -79,13 +79,13 @@ namespace Silownia.Controllers
                 }
                 return View(treningPersonalny);
             }
-           // return HttpNotFound();
+            return HttpNotFound();
         }
 
         // GET: TreningPersonalny/Create
         public ActionResult Create(long? id)
         {
-          //  if (Session["User"] != null)
+            if (Session["User"] != null)
             {
                 ViewBag.TrenerID = new SelectList(db.Trenerzy, "OsobaID", "imieNazwisko");
                 var a = from Osoby in db.Trenerzy select Osoby;
@@ -106,7 +106,7 @@ namespace Silownia.Controllers
 
                 return View();
             }
-           // return HttpNotFound();
+            return HttpNotFound();
         }
 
         // POST: TreningPersonalny/Create
@@ -115,7 +115,7 @@ namespace Silownia.Controllers
         [HttpPost]
         public ActionResult Create([Bind(Include = "TreningID,TreningStart,CzasTrwania,TrenerID")] long? id, TreningPersonalny treningPersonalny)
         {
-         //   if (Session["User"] != null)
+            if (Session["User"] != null)
             {
                 ViewBag.TrenerID = new SelectList(db.Trenerzy, "OsobaID", "imieNazwisko", treningPersonalny.TrenerID);
 
@@ -144,7 +144,7 @@ namespace Silownia.Controllers
                 }
                 return View(treningPersonalny);
             }
-         //   return HttpNotFound();
+            return HttpNotFound();
         }
 
         bool aktywneZajecia(long? klientID, DateTime dataOd)
@@ -181,7 +181,7 @@ namespace Silownia.Controllers
         // GET: TreningPersonalny/Edit/5
         public ActionResult Edit(long? id)
         {
-          //  if (Session["User"] != null)
+            if (Session["User"] != null)
             {
                 if (id == null)
                 {
@@ -195,7 +195,7 @@ namespace Silownia.Controllers
                 ViewBag.MasazystaID = new SelectList(db.Trenerzy, "OsobaID", "imieNazwisko", treningPersonalny.TrenerID);
                 return View(treningPersonalny);
             }
-         //   return HttpNotFound();
+            return HttpNotFound();
         }
 
         // POST: TreningPersonalny/Edit/5
@@ -204,7 +204,7 @@ namespace Silownia.Controllers
         [HttpPost]
         public ActionResult Edit([Bind(Include = "TreningID,TreningStart,CzasTrwania,TrenerID")] TreningPersonalny treningPersonalny)
         {
-           // if (Session["User"] != null)
+            if (Session["User"] != null)
             {
                 if (ModelState.IsValid)
                 {
@@ -215,13 +215,13 @@ namespace Silownia.Controllers
                 ViewBag.TrenerID = new SelectList(db.Trenerzy, "OsobaID", "imieNazwisko", treningPersonalny.TrenerID);
                 return View(treningPersonalny);
             }
-        //    return HttpNotFound();
+            return HttpNotFound();
         }
 
         // GET: TreningPersonalny/Delete/5
         public ActionResult Delete(int? id)
         {
-          //  if (Session["User"] != null)
+            if (Session["User"] != null)
             {
                 if (id == null)
                 {
@@ -234,21 +234,21 @@ namespace Silownia.Controllers
                 }
                 return View(treningPersonalny);
             }
-         //   return HttpNotFound();
+            return HttpNotFound();
         }
 
         // POST: TreningPersonalny/Delete/5
         [HttpPost, ActionName("Delete")]
         public ActionResult DeleteConfirmed(int id)
         {
-         //   if (Session["User"] != null)
+            if (Session["User"] != null)
             {
                 TreningPersonalny treningPersonalny = db.TreningiPersonalne.Find(id);
                 db.TreningiPersonalne.Remove(treningPersonalny);
                 db.SaveChanges();
                 return RedirectToAction("Index", new { akcja = AkcjaEnumTrening.UsunietoTrening });
             }
-         //   return HttpNotFound();
+            return HttpNotFound();
         }
 
         protected override void Dispose(bool disposing)
